2025/requirements.txt
torch
numpy
pandas
scikit-learn
transformers
datasets
evaluate
accelerate
fastapi
uvicorn
python-multipart
firecrawl-py
python-dotenvnltk 
